move:#87 fetchRegularContactsInfinite 를 service.ts로 이동
