+++ src/app/api/supabase/service.ts
@@ -1,4 +1,4 @@
-import { ContactDetailType, ContactItemType, ContactWithPlansDetailType } from '@/types/contacts';
+import { ContactDetailType, ContactItemType, ContactWithPlansDetailType, fetchRegularInfiniteContacts } from '@/types/contacts';
 import { supabase } from '@/app/api/supabase/client';
 import { InsertNewPlansType, PlansType } from '@/types/plans';
 import { CONTACTS, PLANS, USERS } from '@/constants/supabaseTable';
@@ -358,4 +358,26 @@ export const fetchPinnedContacts = async (
 
   if (error) throw error
   return data || []
+}
+
+export const fetchRegularContactsInfinite = async (
+  userId: string,
+  pageParam = 0,
+  limit = 10
+): Promise<fetchRegularInfiniteContacts> => {
+  const from = pageParam * limit
+  const to = from + limit - 1
+
+  const { data, error } = await supabase
+    .from(CONTACTS)
+    .select('contacts_id, name, relationship_level, contacts_profile_img, is_pinned')
+    .eq('user_id', userId)
+    .eq('is_pinned', false)
+    .order('name', { ascending: true })
+    .range(from, to)
+
+  if (error) throw error
+
+  const nextPage = data.length === limit ? pageParam + 1 : undefined
+  return { contacts: data, nextPage }
 }
